requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
ccxt==3.1.54
notifiers==1.3.3
pyTelegramBotAPI==4.14.0 